comparateur_prix/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-filter==24.3
celery==5.3.4
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
python-dotenv==1.0.0
psycopg[binary]==3.2.1
django-cors-headers==4.3.1
django-redis==5.4.0
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
django-otp==1.6.1
qrcode==7.4.2
Pillow==10.4.0
PyJWT==2.9.0
cryptography==43.0.0
elasticsearch==8.12.1
pyzbar==0.1.9
pytesseract==0.3.10
Unidecode==1.3.8
social-auth-app-django==5.4.1

# Graph analytics
networkx==3.2.1
python-louvain==0.16